+++ docs/changelog.rst
@@ -14,6 +14,7 @@ Changes in 0.10.6
 - Add support for mocking MongoEngine based on mongomock. #1151
 - Fixed not being able to run tests on Windows. #1153
 - Allow creation of sparse compound indexes. #1114
+- count on ListField of EmbeddedDocumentField fails. #1187
 
 Changes in 0.10.5
 =================